% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImportTemplates.R
\name{create_import_template}
\alias{create_import_template}
\alias{create_import_template_from_mapping}
\title{Create NASIS XLSX Import Files}
\usage{
create_import_template(
  .data,
  file,
  template_name,
  columns,
  sheet,
  template_version = "1.0",
  delimeter = "<delimiter|||>"
)

create_import_template_from_mapping(
  .data = NULL,
  coltype = c("logical", "physical"),
  file,
  template_name,
  sheet,
  ...
)
}
\arguments{
\item{.data}{Optional: data.frame containing NASIS column physical or logical names to be used to populate the template. Default is an empty template.}

\item{file}{output file name}

\item{template_name}{template name}

\item{columns}{columns in template}

\item{sheet}{XLSX sheet name}

\item{template_version}{template version; default: \code{"1.0"}}

\item{delimeter}{used internally for creating matrix representation for writing to file. Default: \code{"<delimiter|||>"}}

\item{coltype}{Either \code{"physical"} for physical column names, or \code{"logical"} (default) for logical column names (default for worksheet import maps).}

\item{...}{Additional arguments passed to \code{create_import_template()}}
}
\value{
writes XLSX or CSV file
}
\description{
Create NASIS XLSX Import Files
}
\details{
Column names containing \code{"_"} are converted to \code{" "}
}
