% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedonPC.R
\name{create_PedonPC_NASIS_import}
\alias{create_PedonPC_NASIS_import}
\title{Create a NASIS Flat File for importing Pedon Records}
\usage{
create_PedonPC_NASIS_import(
  tables = NULL,
  filename,
  dsn = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{tables}{default \code{NULL} creates an empty template; alternately a list containing data.frames with target table names and new data to write to file.}

\item{filename}{path to output file containing templates for importing each table associated with the NASIS Pedon object}

\item{dsn}{optional path to SQLite Database or an open DBIConnection to a database containing NASIS pedon tables (not yet implemented)}

\item{overwrite}{default \code{FALSE} prevents overwriting \code{filename} if it exists}
}
\value{
a character vector (invisibly) containing the lines/content written out to \code{filename}
}
\description{
Create a NASIS Flat File for importing Pedon Records
}
\author{
Jay Skovlin, Andrew G. Brown
}
