% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSURGO.R
\name{createSSURGO}
\alias{createSSURGO}
\title{Tools for reading NASIS exports}
\usage{
createSSURGO(dsn, output_path = NULL, overwrite = FALSE, append = FALSE)
}
\arguments{
\item{dsn}{Path to \verb{/tabular/} folder of a NASIS export or product available from Web Soil Survey.}

\item{output_path}{Path to output SQLite file or an existing \code{DBIConnection}}

\item{overwrite}{Default: \code{FALSE}; passed to \code{DBI::dbWriteTable()}}

\item{append}{Default: \code{FALSE}; passed to \code{DBI::dbWriteTable()}}
}
\value{
an invisible named list of data.frame. If \code{output_path} is not \code{NULL} logical; \code{TRUE} all tables are written to path
}
\description{
Read from a tabular SSURGO export folder (.txt) files, and convert to SQLite database.
}
