% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImportTemplates.R
\name{create_ESD_ecosites_import}
\alias{create_ESD_ecosites_import}
\alias{create_ESD_notes_import}
\title{Create Ecosite / Ecosite Note Import Files}
\usage{
create_ESD_ecosites_import(file, coiids, ecositeids)

create_ESD_notes_import(file, coiids, author, notes)
}
\arguments{
\item{file}{output file name (either .XLSX or .CSV)}

\item{coiids}{vector of component IDs}

\item{ecositeids}{vector of ecological site IDs}

\item{author}{author of note}

\item{notes}{note content}
}
\value{
writes XLSX or CSV file
}
\description{
Create Ecosite / Ecosite Note Import Files
}
