% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosite.R
\name{ecositeid_to_name}
\alias{ecositeid_to_name}
\title{Convert Ecological Site ID to Ecological Site Name using NASIS}
\usage{
ecositeid_to_name(x)
}
\arguments{
\item{x}{character vector of Ecological Site IDs}
}
\description{
Convert Ecological Site ID to Ecological Site Name using NASIS
}
\details{
Uses contents of local NASIS \code{ecologicalsite} table
}
