% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationStatus.R
\name{get_correlation_info}
\alias{get_correlation_info}
\title{Get Correlated Mapunit and Component Information}
\usage{
get_correlation_info(x = NULL, x.col = "dmuiid", additional = FALSE)
}
\arguments{
\item{x}{vector of record IDs (Default \code{NULL} is entire local database)}

\item{x.col}{physical column name containing record IDs (e.g. liid, muiid, dmuiid, coiid, ecositeiid). Default \code{"dmuiid"}}

\item{additional}{include additional mapunits? Default \code{FALSE}}
}
\value{
a data.frame containing component-level information and IDs for all parent tables up to legend. Results will contain component information from all related mapunits.
}
\description{
Get the correlated mapunit status for a vector of IDs. These IDs could be from area, legend, legendmapunit, mapunit correlation, data mapunit, component, component ecological site or ecological site.
}
