% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NASIS_metadata.R
\name{get_NASIS_metadata}
\alias{get_NASIS_metadata}
\title{Get NASIS metadata entries for specific domains or choices}
\usage{
get_NASIS_metadata(x, what = "ColumnPhysicalName", dsn = NULL)
}
\arguments{
\item{x}{Vector to match in NASIS metadata}

\item{what}{Column to match \code{x} against. Default "ColumnPhysicalName"; alternate options include \code{"DomainID"}, \code{"DomainName"}, \code{"DomainRanked"}, \code{"DisplayLabel"}, \code{"ChoiceSequence"}, \code{"ChoiceValue"}, \code{"ChoiceName"}, \code{"ChoiceLabel"}, \code{"ChoiceObsolete"}, \code{"ChoiceDescription"}, \code{"ColumnLogicalName"}}

\item{dsn}{passed to soilDB:::.get_NASIS_metadata() for alternate NASIS database connection types (path to SQLite file or open DBIConnection)}
}
\value{
a data.frame containing selected NASIS metadata sorted first on \code{DomainID} and then on \code{ChoiceSequence}
}
\description{
Get NASIS metadata entries for specific domains or choices
}
\examples{
get_NASIS_metadata("texcl")
}
