% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_tools.R
\name{get_projectmapunits_by_uprojectid}
\alias{get_projectmapunits_by_uprojectid}
\title{get_projectmapunits_by_uprojectid}
\usage{
get_projectmapunits_by_uprojectid(
  uprojectid,
  areasymbol = NULL,
  dsn = soilDB::NASIS()
)
}
\arguments{
\item{uprojectid}{vector of user project IDs}

\item{areasymbol}{Soil Survey Legend areasymbol to include in WHERE clause; Default: \code{NULL} is no constraint on areasymbol}

\item{dsn}{data source name, default is \code{soilDB::NASIS()} \emph{OdbcConnection}}
}
\value{
data.frame result
}
\description{
get_projectmapunits_by_uprojectid
}
