% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO-export-Interpretations.R
\name{get_SSURGO_interp_reasons_by_mrulename}
\alias{get_SSURGO_interp_reasons_by_mrulename}
\title{Get Interpretation Rating "Reasons" from SSURGO \code{cointerp} table}
\usage{
get_SSURGO_interp_reasons_by_mrulename(dsn, mrulename, n = 2)
}
\arguments{
\item{dsn}{A DBIConnection}

\item{mrulename}{Rule name of interpretation}

\item{n}{Number of reasons to return}
}
\value{
A \code{data.frame} containing columns: "lmapunitiid", "coiid", "mrulename", "cokeyref", "Reasons", "liid", "muiid", "corriid", "dmuiid", "areasymbol", "musym", "compname", "comppct_r", "interphr", "interphrc","mukey"
}
\description{
Get Interpretation Rating "Reasons" from SSURGO \code{cointerp} table
}
\seealso{
\code{\link[=get_SSURGO_cointerp]{get_SSURGO_cointerp()}} \code{\link[=get_SSURGO_component_keys]{get_SSURGO_component_keys()}}
}
