% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupTexture.R
\name{lookupTexture}
\alias{lookupTexture}
\alias{lookupTextureModifier}
\title{Lookup Soil Texture Abbreviation Codes or Names}
\usage{
lookupTexture(x, what = "names", dsn = NULL)

lookupTextureModifier(x, what = "names", dsn = NULL)
}
\arguments{
\item{x}{a vector of soil texture/texture modifier abbreviation codes or names}

\item{what}{either \code{"names"} (default) or \code{"codes"}. The opposite of what \code{x} is.}

\item{dsn}{passed to \code{get_NASIS_metadata()}}
}
\value{
a vector of soil texture abbreviation codes or names depending on whether input values are matched in the lookup table and the value of \code{what}. Unmatched values in \code{x} return \code{NA}.
}
\description{
Lookup Soil Texture Abbreviation Codes or Names
}
\examples{
x <- lookupTexture(c("loam", "bar", "sandy loam"), what = "codes")
x
lookupTexture(x, what = "names")

lookupTextureModifier("extremely bouldery")
}
