% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImportMapping.R
\name{read_import_mapping}
\alias{read_import_mapping}
\title{Read Worksheet Import Mapping XML

Reads a worksheet import map, identifies physical column names and associated columns for worksheets.}
\usage{
read_import_mapping(wsname)
}
\arguments{
\item{wsname}{Worksheet Name}
}
\value{
data.frame
}
\description{
Read Worksheet Import Mapping XML

Reads a worksheet import map, identifies physical column names and associated columns for worksheets.
}
