% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapColumns.R
\name{remapColumns}
\alias{remapColumns}
\title{Map columns of a SoilProfileCollection or data.frame to a custom schema}
\usage{
remapColumns(x, .map)
}
\arguments{
\item{x}{a SoilProfileCollection or horizon-level data.frame}

\item{.map}{a named, nested list where each sub-list defines the columns associated with a new table in terms of existing columns in \code{x}. The basic format of the .map argument follows: \code{list(new_table_name = list(new_column_name = 'old_column_name'))}}
}
\value{
a named list of data.frame objects (equal in length to \code{.map}) containing data from \code{x} "remapped" to the schema specified by \code{.map}.
}
\description{
Map columns of a SoilProfileCollection or data.frame to a custom schema
}
