% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO-export-Interpretations.R
\name{get_SSURGO_cointerp}
\alias{get_SSURGO_cointerp}
\alias{get_SSURGO_component_keys}
\title{Get SSURGO Component Interpretations summaries}
\usage{
get_SSURGO_cointerp(dsn, mrulename = NULL, ruledepth = 0)

get_SSURGO_component_keys(dsn, mrulename = NULL, ruledepth = 0)
}
\arguments{
\item{dsn}{A DBIConnection}

\item{mrulename}{Filter on rule name(s) of interpretation}

\item{ruledepth}{Filter rule depth (default \code{ruledepth = 0})}
}
\description{
Get SSURGO Component Interpretations summaries
}
