% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImportTemplates.R
\name{create_import_template}
\alias{create_import_template}
\title{Create NASIS import template}
\usage{
create_import_template(
  .data,
  file,
  template_name,
  columns,
  template_version = "1.0"
)
}
\arguments{
\item{.data}{a data.frame containing source data}

\item{file}{output file name}

\item{template_name}{template name}

\item{columns}{columns in template}

\item{template_version}{template version; default: \code{"1.0"}}
}
\value{
writes XLSX or CSV file
}
\description{
Create NASIS import template
}
\details{
Column names containing \code{"_"} are converted to \code{" "}
}
