% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosite.R
\name{ecositeid_to_name}
\alias{ecositeid_to_name}
\title{Convert Ecological Site ID to Ecological Site Name}
\usage{
ecositeid_to_name(x)
}
\arguments{
\item{x}{A vector of Ecological Site IDs}
}
\description{
Convert Ecological Site ID to Ecological Site Name
}
\details{
Uses contents of NASIS \code{ecologicalsite} table
}
