% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_tools.R
\name{get_lmapunit_by_areasymbol}
\alias{get_lmapunit_by_areasymbol}
\title{get_lmapunit_by_areasymbol}
\usage{
get_lmapunit_by_areasymbol(areasymbol, dsn = soilDB::NASIS())
}
\arguments{
\item{areasymbol}{vector of areasymbols}

\item{dsn}{data source name, default is \code{soilDB::NASIS()} \emph{OdbcConnection}}
}
\value{
data.frame result
}
\description{
get_lmapunit_by_areasymbol
}
