% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_tools.R
\name{get_projectmapunits_by_uprojectid}
\alias{get_projectmapunits_by_uprojectid}
\title{get_projectmapunits_by_uprojectid}
\usage{
get_projectmapunits_by_uprojectid(dsn = soilDB::NASIS(), uprojectid)
}
\arguments{
\item{dsn}{data source name, default is \code{soilDB::NASIS()} \emph{OdbcConnection}}

\item{uprojectid}{vector of user project IDs}
}
\value{
data.frame result
}
\description{
get_projectmapunits_by_uprojectid
}
